requirements-devel.txt
# requirements for a development environment
-e .[devel]
datalad-metalad
